Add Debian packaging.
Add necessary debian/* files and a new makefile goal, "deb",
for building a Debian package containing the Logstash Filter
Verifier binary.

To avoid having to update debian/changelog every time you want
to build a package with a meaningful version number we use
"git describe" to generate a version number and automatically
add a changelog entry if the automatically generated version
doesn't match the most recent entry (since it's the most recent
changelog entry that becomes the version of the built package).
This will in most cases create a dirty debian/changelog every
time a package is built but that's a better problem to have.