Add --version flag.
The version number in the --version output comes from version.go,
which we generate in the makefile using "git describe" (same as
for the Debian package version).

The makefile-based generation of version.go means we have to build
with GNU make, at least to create a bootstrap copy of version.go
(lest one types the file by hand).