Force single filter worker to avoid pipeline races.
Starting with Logstash 2.0, the default number of filter workers
(the -w/--filterworkers startup option) is equal to half the number
of CPU cores. This can result in failing tests because the ordering
of events coming out of the pipeline is suddenly no longer given,
so we limit the number of workers to one to match the pre-2.0
default.