Don't use ExistingFile() for validation of --logstash-path.
As tempting as it is to use ExistingFile() for flag definitions
and get automatic validation, its validation kicks in
unconditionally and too early for flags with default values,
so not even --version will work if you don't have Logstash
installed as /opt/logstash/bin/logstash.

Let's disable the --logstash-path validation altogether until
we've figured out how to do this better. Having the validation
done by the flag parser is really just an optimization.