Start using GPM for managing dependencies.
To secure that we can build consistent binaries (indeed, that we
can build binaries at all) we start using the GPM tool to lock our
dependencies to tags or SHA-1s. Of course, if the upstream gits are
rewound or deleted we're still out of luck since we're not vendoring
the third party packages.