Pass Logstash a dir with the config files, not multiple --config opts.
It was a mistake to assume that Logstash accepts multiple --config
options as a way of passing multiple configuration files. Only one
of the configuration files is actually picked up, i.e. using Logstash
Filter Verifier only worked if you had all your filters in the same
file.

Instead we copy all configuration files into a temporary directory
that we pass to Logstash.