Use one-based numbering of temp dirs for result files.
Previously the directories storing the "actual" and "expected" files
were using the zero-based event index, but that led to confusing
messages like this:

    Comparing message 1 of foo.json...
    Comparing message 2 of foo.json...
    --- /tmp/572340849/http-access.json/1/expected	2016-08-01 ...
    +++ /tmp/572340849/http-access.json/1/actual	2016-08-01 ...

In this example it's ambiguous whether it's the first or the second
event in foo.json that failed the test. If we use a one-based index
things are clear.