Add support for value of object type
Nested fields (values of object type) are addressed in Logstash with
field reference notation [1]. The nested field names are put into square
brackets [], starting on the top level. If a top level field is addressed,
the square brackets might be omited.

Currently values of object type are not allowed in arrays.

Fixes #12

[1]: https://www.elastic.co/guide/en/logstash/current/event-dependent-configuration.html#logstash-config-field-references