Parallel execution of multiple test cases
Implements the usage of logstash-input-unix as input, which
allows to use a single Logstash instance for the execution
of multiple test case files, which greatly shortens the
duration for the test execution.

To enable this feature, run logstash-filter-verifier with the
command line flage `--sockets`.
This feature does only work on systems, which do support
unix domain sockets.

The logstash-input-unix plugins adds the two fields `host` and
`path`. Therefore these fields should be ignored in the test cases.

This implementation does need at least Go 1.8beta1, as the function
net.SetUnlinkOnClose is used, which is introduced with Go 1.8.

Fixes #14