Added testcases section to the test case file
The test case file now supports an additional section
called `testcases`, which consists of an input line and the
expected event, after processed by Logstash.

This allows for clean test case files, as the actual input is
just next to the expected output, which makes working with
the test case files much easier.