parallel_process: Delete "path" fields that contain socket path.
Logstash's unix input adds a "path" field containing the socket path,
which screws up the test results unless the filters happen to delete
that field. We can't unconditionally delete that field ourselves because
the input JSON payload could contain a "path" field that we can't touch,
but we can safely delete the field if its contents is equal to the
socket path.