Preserve PATH variable by default.
For some reason Logstash's bootstrap script has been able to find the
JVM even though the PATH variable was unset, but this wasn't true when
running Logstash 5.4 via the docker.elastic.co/logstash/logstash:5.4.0
Docker image, resulting in a "Could not find any executable java
binary" error message.

It's not clear whether this is a general 5.x or 5.4 problem or if
it's related to the Docker image, but either way it's reasonable to
preserve the PATH variable by default. It's not likely to result in
non-hermetic tests, and if any such concerns arise it's easy to set
the variable to a fixed value or unset it (if the JVM can be found
anyway).