Introduce --logstash-version and use it to pass correct log option.
The new --logstash-version flag chooses which version of Logstash
that's being started, allowing LFV to adapt itself. If the flag is
omitted we'll run "logstash --version" to auto-detect the version.

Over time we'll use this for lots of things but the first one out
is Logstash's -l flag that starting with Logstash 5.0 needs to be
a directory instead of a file.

Fixes #41