Search for Logstash executable in multiple places.
Previously the --logstash-path flag has been a singular value, but since
the Logstash installation directory in RPM and Debian packages changed
in Logstash 5.0 it's desirable to have multiple values.

We change the meaning of the flag to append to the list of paths that
are being tried to locate an executable instead of overwriting the
previously set value.