logstash: Use single temporary directory per invocation.
Support for Logstash 5.0+ is going to require us to add additional
temporary directories for e.g. configuration and data, so let's
shift to a single tempdir per invocation.

This required logstash.getConfigFileDir() to no longer return the
path to a new tempdir but rather store the files in the directory
that's passed as an argument.