logstash: Refactor test table in invocation_test.go
To make it easier to introduce additional checks for the computed
Logstash options in the invocation, introduce a simple option
parser and pass a map of all options to the test function defined
in the test table. Then it'll be up to each test in the table what
aspects of the options to test.