logstash: Reject testcases with "fields" key set to null
Previously a testcase file containing

    {
        "fields": null,
        ...
    }

would be accepted but would result in a panic during execution if
--sockets was used since TestCaseSet.InputFields would be nil. We
now reject null values to avoid this. If no fields should be set
users should omit the key or set it to an empty object. Fixes #54.

This bug was similar to the one fixed in commit db42e83, but that
commit fixed the case of when TestCaseSet.InputFields became null
because the key was omitted rather than explicitly set to null.