Migrate to Go modules
We replace the use of GPM for package management with Go's built-in
modules. Apart from converting the Godeps file to go.mod we've updated
the makefile to not store binaries in $GOPATH/bin but in $GOBIN, and
we set the latter to ./bin. We also had to change the import path for
github.com/Masterminds/semver to add a /v3 suffix since the version
of that package is 3.0.1.