Step up to Go 1.13.3 and drop Debian packaging support
Let's keep up with the Go releases to be able to use modules.
Since no stock Debian packages of Go 1.13 are available (yet)
Debian packaging of LFV is dropped. The Debian packaging has
been a hassle on more than one occasion and I'd rather not
spend time on that anymore.