Replace go-metalinter with golangci-lint
golangci-lint is much faster and more configurable. With the current
configuration the LFV code has zero violations so we can actually add
"make check" to the Travis configuration. Follow-up commits will enable
more checks and fix the ensuing violations.

We fetch the tool via curl (as is recommended) instead of via Godeps.
There's a limitation in that changing the GOLANGCI_LINT_VERSION make
variable doesn't force a download of a different version of
golangci-lint. Having make targets depend on variable values is
totally doable but cumbersome. We'll leave it like this for now.