Use govvv to "git describe"-stamp binaries
Generating a version.go file was fine when we could assume that LFV
was always built via the makefile, but after the transition to modules
that's less likely. The govvv tool solves this for us, although if you
import LFV as a module the main.GitSummary variable won't be populated
and will read "(unknown)". If that's an important use case it's possible
to read the module version via debug.ReadBuildInfo().