Include testcase filename in error message
In Unix socket mode it wasn't always clear which testcase file
failed. If the failure was a mismatch in one or more events
things were fine, but if e.g. the event count was wrong all you
got was this:

    Testcase failed, continuing with the rest: Expected 4 event(s),
    got 0 instead.

We now include the testcase filename in this error message.