Add "checktidy" make goal for checking "go mod tidy" cleanliness
We don't want cruft laying around in go.{mod,sum} and the new
"checktidy" make goal makes it easy to check. Since the command
could modify the working tree and produces false positives if
run from a tree where go.{mod,sum} already is dirty we don't run
it as part of the old "check" goal. However, include it in the
Travis build.