Make sure observers have time to complete their processing
There was a race condition when lfvobserver.TestExecutionEnd{} had
been pushed to the summary observer since we never waited for it
to write its output before shutting down. This was okay most of
the time but occasionally the summary output lost the race.

Introduce an lfvobserver.Interface interface that observers can
adhere to, including a Finalize method that waits for the final
event to have been processed.

Fixes #92