Fix automatic IDs
Automatically adding the IDs did not work before, because during the
creation of the ZIP file, that is passed to the daemon, the Logstash
config files have been read from disk again. With this, the original
config has been passed along, which does not contain the required IDs.

To prevent this, the validation step for the Logstash config is now
part of the creation of the ZIP file.

Fixes: #131