Change default loglevel from WARNING to INFO
With WARNING as the default loglevel the daemon wouldn't emit any useful
status messages to let you know it has actually started up and is ready
to be used. INFO will give users a few more messages overall but there's
no dramatic increase and it's not likely to be regarded as spammy.