Use standard gopkg.in/yaml.v2 package for tests
The use of github.com/mikefarah/yaml/v2 was motivated because the
default type for maps was `map[interface{}]interface{}` instead of
`map[string]interface{}`. This is still the case, but this does not seem
to be a problem anymore. The fork used was forcibly synced with the
original package and we now get:

verifying github.com/mikefarah/yaml/v2@v2.4.0/go.mod: checksum mismatch
        downloaded: h1:RDklbk79AGWmwhnvt/jBztapEOGDOx6ZbXqjP6csGnQ=
        go.sum:     h1:ahVqZF4n1W4NqwvVnZzC4es67xsW9uR/RRf2RRxieJU=