Remove support for old test case files
Originally, the "input" and "expected" keys were at the top level
of the test case files, and since the introduction of the "testcases"
key both have been supported. Keeping the old keys is only confusing,
so now that we're stepping the major version we should seize the
opportunity to drop support for the old format.

Rather than just drop them entirely and let the YAML package report
the error, we capture any values given these keys and provide the user
with a useful error message.